hana_performance_reports/requirements.txt
hdbcli>=2.19.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
